Secured all the other AssociateController methods, except mock data generation.

+++ StagingManagementSystem/src/main/java/com/revature/controllers/rest/AssociateControllerImpl.java
@@ -35,16 +35,24 @@ public class AssociateControllerImpl {
 	}
 
 	@PostMapping
-	public void addAssociate(@RequestBody Associate associate) {
+	public ResponseEntity addAssociate(@RequestBody Associate associate, HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
 		associateService.add(associate);
+		return ResponseEntity.ok(null);
 	}
 	
 	 @PostMapping("/add/all")
-	  public void addAssociates(@RequestBody Set<Associate> associates) {
-	   for(Associate a : associates){
-       associateService.add(a);       
-	   }
-	  }
+	 public ResponseEntity addAssociates(@RequestBody Set<Associate> associates, HttpSession session) {
+		 if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		 }
+		 for(Associate a : associates){
+			 associateService.add(a);
+		 }
+		 return ResponseEntity.ok(null);
+	 }
 	 
 	 @GetMapping("/generate/mock-data")
 	 public void generateAssociateMockDate(){
@@ -52,8 +60,12 @@ public class AssociateControllerImpl {
 	 }
 
 	@DeleteMapping
-	public void deleteAssociate(@RequestBody Associate associate) {
-		associateService.delete(associate);
+	public ResponseEntity deleteAssociate(@RequestBody Associate associate, HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+	 	associateService.delete(associate);
+		return ResponseEntity.ok(null);
 	}
 
 	@PutMapping
@@ -73,12 +85,18 @@ public class AssociateControllerImpl {
 	}
 
 	@GetMapping("/{id}")
-	public Associate getAssociate(@PathVariable long id) {
-		return associateService.getById(id);
+	public ResponseEntity<Associate> getAssociate(@PathVariable long id, HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+	 	return ResponseEntity.ok(associateService.getById(id));
 	}
 
 	@GetMapping("/all")
-	public Set<Associate> findById() {
-		return associateService.getAll();
+	public ResponseEntity<Set<Associate>> getAllAssociates(HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+	 	return ResponseEntity.ok(associateService.getAll());
 	}
 }
